microservice/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# File type detection
python-magic==0.4.27

# Async support
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Production optimizations
gunicorn==21.2.0

# Optional Redis support
redis==5.0.1

# MarkItDown will be installed from local package